refactor: Response, Request DTO 에 프로토콜 채택
